shop-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
python-dotenv==1.0.0 